I'm officially over the yellowish background
While I thought it was nice at one time and had perhaps the feel of a
well loved old book, I think it's begun to hold me back. I longed to
once again let my eyes roam the crisp expanses of a background that
didn't steal the stage from the other players. And so today, we are back
to #fff.